No issue: rename web pkg to iwebview.
Calling it "web" makes the difference between that and the "browser"
pkg ambiguous.
