+++ app/src/main/java/org/mozilla/focus/iwebview/IWebView.kt
@@ -3,7 +3,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package org.mozilla.focus.web
+package org.mozilla.focus.iwebview
 
 import android.graphics.Bitmap
 import android.support.annotation.CheckResult
