+++ app/src/main/java/org/mozilla/focus/session/SessionCallbackProxy.java
@@ -8,7 +8,7 @@ import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 import android.view.View;
 
-import org.mozilla.focus.web.IWebView;
+import org.mozilla.focus.iwebview.IWebView;
 
 public class SessionCallbackProxy implements IWebView.Callback {
     /* package */ static final int MINIMUM_PROGRESS = 5;
