+++ app/src/main/res/layout/fragment_browser.xml
@@ -23,13 +23,13 @@
             android:orientation="vertical"
             android:clipChildren="false">
 
-            <org.mozilla.focus.web.IWebView
+            <org.mozilla.focus.iwebview.IWebView
                 android:id="@+id/webview"
                 android:layout_width="match_parent"
                 android:layout_height="match_parent"
                 android:scrollbarThumbVertical="@drawable/scrollbar_thumb">
                 <requestFocus />
-            </org.mozilla.focus.web.IWebView>
+            </org.mozilla.focus.iwebview.IWebView>
 
         </org.mozilla.focus.widget.ResizableKeyboardCoordinatorLayout>
     </LinearLayout>
