+++ app/src/main/res/layout/fragment_new_settings.xml
@@ -152,11 +152,11 @@
         </LinearLayout>
 
     </LinearLayout>
-    <org.mozilla.focus.web.IWebView
+    <org.mozilla.focus.iwebview.IWebView
         android:id="@+id/settingsWebView"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:scrollbarThumbVertical="@drawable/scrollbar_thumb"
         android:visibility="gone">
-    </org.mozilla.focus.web.IWebView>
+    </org.mozilla.focus.iwebview.IWebView>
 </LinearLayout>
